+++ ex02/pokedex/src/components/NavBar.vue
@@ -13,57 +13,21 @@
       </div>
       <v-spacer></v-spacer>
       <search-menu></search-menu>
-      <v-menu offset-y>
-        <template v-slot:activator="{ on, attrs }">
-          <v-btn
-            color="amber lighten-1"
-            class="blue-grey--text"
-            rounded
-            v-bind="attrs"
-            v-on="on"
-          >
-            {{ langDisplayName }}
-            <v-icon>mdi-menu-down</v-icon>
-          </v-btn>
-        </template>
-        <v-list>
-          <v-subheader>LANGUAGE</v-subheader>
-          <v-list-item-group v-model="selectedLanguageIndex" color="primary">
-            <v-list-item
-              v-for="[langName, displayName] in Array.from(langNameMap)"
-              :key="langName"
-            >
-              <v-list-item-content @click="changeLanguage(langName)">{{
-                displayName
-              }}</v-list-item-content>
-            </v-list-item>
-          </v-list-item-group>
-        </v-list>
-      </v-menu>
+      <lang-menu></lang-menu>
     </v-app-bar>
   </div>
 </template>
 
 <script lang="ts">
 import { Component, Vue } from "vue-property-decorator";
-import { Language, languageNameMap } from "@/models/language";
 import SearchMenu from "@/components/SearchMenu.vue";
+import LangMenu from "@/components/LangMenu.vue";
 
 @Component({
   components: {
     SearchMenu,
+    LangMenu,
   },
 })
-export default class NavBar extends Vue {
-  private langNameMap = languageNameMap;
-  private selectedLanguageIndex = 0;
-
-  get langDisplayName() {
-    return this.langNameMap.get(this.$store.state.language);
-  }
-
-  changeLanguage(lang: Language) {
-    this.$store.commit("changeLanguage", lang);
-  }
-}
+export default class NavBar extends Vue {}
 </script>
